CompViewer downloads the required libs on first run.
If a internet card is installed, and the http access is enabled on the server we will try to download the required libs on first run.  We also check for updates on each run, so users no longer have to check the repo for updates manually.  Also added is "compviewer-version.txt" that has the latest version number that we check against.  We also now download a updated ComponentInfo file every run, if we are able to, this way you are not required to do a full update to just update Component Info.  Any changes to gml.lua, default.gss, CompViewer.lua, gfxbuffer.lua, or colorutils.lua will require you to bump the version number in the text file. Decimals are valid eg 1.0 -> 1.1 -> 1.2 etc.